Move stop string processor finalization logic out of generate.py
